Begin adding basic scripting
Started adding scripting as a module, then decided it would be better
for it to be a class.
(There's a little bit of re-work to be done to change it to a class.)

Very basic scripting "language" that will allow file/folder manipulation
like writing to a file, creating a file/folder, as well as pop up a
small dialog box for input of whatever you'd like. Also plan to adda
resolution picker.
Set and reference basic variables, including the screen width and
height.