Big Scripting Changes
I'd stumbled upon a small scripting language on PlanetSourceCode and
I've been adding to it and trying to improve it for inclusion in the
LanParty client. I'm only uploading the class and the debug form itself
as of right now, but I hope to also add a simple code editor and get
everything nicely packaged.